+++ changelog.md
@@ -1,4 +1,5 @@
-# AlchemyLib 1.20.1-1.0.28 RELEASE
+# AlchemyLib 1.20.1-1.0.29 RELEASE
 
 Changes:
-- Add SideConfig state boolean to ProcessingBlockEntity to allow for better user experience opening and closing side config screens.
+- Move SideModeButton out of Alchemistry and into AlchemyLib.
+- Refactor RecipeSelectorButton such that it won't open the screen if the side mode screen is already open.